Changed types again because it didn't run on Windows.

+++ bluecontroller/avrdude/avrdude-5.10.bt/bluec.c
@@ -161,7 +161,7 @@ static int bluec_bulk_write(PROGRAMMER * pgm, AVRPART * p, AVRMEM * m,
 {
   int rc = n_bytes; // return code, negative on failure
 
-  const unsigned int bufSize = 1024;
+  const unsigned long bufSize = 1024;
   unsigned char buf[bufSize]; // for receiving a failed verify notification we need some reserve
   //int block_size;
   //int tries;
@@ -203,15 +203,15 @@ static int bluec_bulk_write(PROGRAMMER * pgm, AVRPART * p, AVRMEM * m,
     return rc;
   }
   bulkOptions &= buf[1]; // maybe not all requested options could be enabled, so mask out the ones which aren't active
-  const unsigned int windowSize = buf[2] | buf[3] << 8;
+  const unsigned short windowSize = buf[2] | buf[3] << 8;
   unsigned long writePos = 0; // counting of the written protocol bytes, starts after reading BULK_WRITE_START_ACK
   unsigned long ackPos = 0; // counting of the acknowledged protocol bytes, starts after reading BULK_WRITE_START_ACK
 
-  const unsigned int mtu = 126; // this is what I get for a RFComm commnection on Mac OSX 10.7 TODO: should be read from serial driver after connecting
-  //const unsigned int minimumBlockSize = 64; // don't send smaller packets
-  const unsigned int writeDataOverhead = 6; // the protocol overhead of STK_BULK_WRITE_DATA
-  const unsigned int writeVrfyOverhead = 8; // the protocol overhead of BULK_WRITE_VRFY_ERR
-  unsigned int block_size = mtu - writeDataOverhead; // netto data of one block
+  const unsigned short mtu = 126; // this is what I get for a RFComm commnection on Mac OSX 10.7 TODO: should be read from serial driver after connecting
+  //const unsigned short minimumBlockSize = 64; // don't send smaller packets
+  const unsigned long writeDataOverhead = 6; // the protocol overhead of STK_BULK_WRITE_DATA
+  const unsigned long writeVrfyOverhead = 8; // the protocol overhead of BULK_WRITE_VRFY_ERR
+  unsigned short block_size = mtu - writeDataOverhead; // netto data of one block
   if((block_size + writeDataOverhead) > bufSize)
     block_size = bufSize - writeDataOverhead;
   if(block_size > (windowSize - 2) / 2) // make sure at least two blocks can be traveling at the same time
@@ -241,10 +241,10 @@ static int bluec_bulk_write(PROGRAMMER * pgm, AVRPART * p, AVRMEM * m,
 
       if(remain > 0)
       {
-        unsigned int cursize = block_size;
+        unsigned short cursize = block_size;
         if(cursize > remain)
           cursize = remain;
-        unsigned int crc = 0;
+        unsigned short crc = 0;
         i = 0;
         buf[i++] = Cmnd_STK_BULK_WRITE_DATA;
         buf[i++] = crc & 0xff;
@@ -346,7 +346,7 @@ fprintf(stderr, "ackn(%6d, %6d, %6d, %6d, cmd=0x%02x)\n", writePos, ackPos, byte
             goto bulkReceiveEnd;
           }
           unsigned long vrfyAddr = buf[1] | (buf[2] << 8) | (buf[3] << 16) | (buf[4] << 24);
-          unsigned int curSize = buf[5] | (buf[6] << 8);
+          unsigned short curSize = buf[5] | (buf[6] << 8);
           if ((curSize + writeVrfyOverhead) > bufSize)
           {
             fprintf(stderr, "%s: bluec_bulk_write(): Overflow while reading data of VRFY_ERR\n", progname);
